+++ android/src/main/java/com/nasmedia/admixerssp/NapListenerBridge.kt
@@ -2,7 +2,10 @@ package com.nasmedia.admixerssp
 
 interface NapListenerBridge {
     fun onReceivedAd(adapterName: String, ad: Any)
-    fun onFailedToReceiveAd(ad: Any?, name: String, code: Int, msg: String?)
+    fun onFailedToReceiveAd(code: Int, msg: String?) {
+        onFailedToReceiveAd(null, "", code, msg)
+    }
+    fun onFailedToReceiveAd(ad: Any?, name: String, code: Int, msg: String?) {}
     fun onAdShowFailed(ad: Any?, name: String, code: Int, msg: String?) {}
     fun onAdClicked() {}
     fun onAdClosed() {}